pyproject.toml
[tool.poetry]
name = "cdctrader"
version = "0.1.0"
description = ""
authors = ["Omar Baumgartner"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyyaml = "^6.0.1"
websocket-client = "^1.6.4"
numpy = "^1.26.2"
telethon = "^1.33.1"
networkx = "^3.2.1"
matplotlib = "^3.8.2"
mplfinance = "^0.12.10b0"
sympy = "^1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
